requirements.txt
PyQt6>=6.4.0
Pillow>=9.0.0
opencv-python>=4.7.0
numpy>=1.24.0
